+++ database/migrations/2023_09_23_145230_add_scheduled_release_date_time_column_to_show_episodes_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('show_episodes', function (Blueprint $table) {
+            $table->dateTime('scheduled_release_dateTime')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('show_episodes', function (Blueprint $table) {
+            $table->dropColumn(['scheduled_release_dateTime']);
+        });
+    }
+};