+++ resources/js/Components/Shows/ShowFooter.vue
@@ -4,8 +4,9 @@
         <!-- Paginator -->
         <!--                            <Pagination :links="`#`" class="mt-6"/>-->
         <Link :href="`/teams/${team.slug}`" class="text-blue-500 ml-2 uppercase"> {{ team.name }} ©
-            <span v-if="show.last_release_year" >{{ show.first_release_year }}-{{show.last_release_year}}</span>
-            <span v-if="!show.last_release_year">{{show.first_release_year}}</span>
+            <span v-if="show.last_release_year > 0" >{{ show.first_release_year }}-{{show.last_release_year}}</span>
+            <span v-if="!show.last_release_year && show.first_release_year">{{show.first_release_year}}</span>
+            <span v-if="!show.last_release_year && !show.first_release_year">{{currentYear}}</span>
         </Link>
     </div>
 
@@ -18,6 +19,6 @@ defineProps({
     show: Object,
 })
 
-let thisYear = new Date().getFullYear()
+let currentYear = new Date().getFullYear()
 
 </script>